misc(compass-components): Move ui colors out of components folder: Add a few helpful deps; Fix testing-lib version collision
