+++ package.json
@@ -45,8 +45,12 @@
     "type": "git",
     "url": "git://github.com/mongodb-js/compass.git"
   },
-  "//": "TODO: This is a hack to make @webpack-cli/serve dependency unhoist back to the webpack config package. This can be removed after all packages are using the same webpack config.",
+  "//": "TODO: We keep webpack-cli/serve and testing-library/dom at the root to ",
+  "//": "      work around weird npm workspace hoising issues caused by outdated",
+  "//": "      transient dependencies. This can go away when we update webpack  ",
+  "//": "      and leafygreen to the same version everywhere",
   "devDependencies": {
+    "@testing-library/dom": "^8.11.1",
     "@webpack-cli/serve": "^0.2.0",
     "chalk": "^4.1.1",
     "find-up": "^5.0.0",
