+++ packages/compass-components/package.json
@@ -40,6 +40,7 @@
     "@leafygreen-ui/card": "^5.1.3",
     "@leafygreen-ui/checkbox": "^6.0.5",
     "@leafygreen-ui/confirmation-modal": "^2.2.1",
+    "@leafygreen-ui/emotion": "^4.0.0",
     "@leafygreen-ui/icon": "^11.4.0",
     "@leafygreen-ui/icon-button": "^9.1.6",
     "@leafygreen-ui/leafygreen-provider": "^2.1.2",
@@ -57,7 +58,9 @@
     "@leafygreen-ui/tokens": "^0.5.3",
     "@leafygreen-ui/tooltip": "^6.2.2",
     "@leafygreen-ui/typography": "^8.0.2",
-    "@mongodb-js/compass-logging": "^0.3.0"
+    "@mongodb-js/compass-logging": "^0.3.0",
+    "@react-aria/interactions": "^3.6.0",
+    "@react-aria/visually-hidden": "^3.2.3"
   },
   "peerDependencies": {
     "@emotion/css": "^11.5.0",
@@ -69,7 +72,9 @@
     "@mongodb-js/mocha-config-compass": "^0.4.0",
     "@mongodb-js/prettier-config-compass": "^0.2.0",
     "@mongodb-js/tsconfig-compass": "^0.2.0",
+    "@testing-library/dom": "^8.11.1",
     "@testing-library/react": "^12.0.0",
+    "@testing-library/user-event": "^13.5.0",
     "@types/chai": "^4.2.21",
     "@types/chai-dom": "^0.0.10",
     "@types/mocha": "^9.0.0",
