+++ packages/compass-components/src/index.ts
@@ -28,7 +28,7 @@ export {
 export { Menu, MenuSeparator, MenuItem } from '@leafygreen-ui/menu';
 export { default as Modal } from '@leafygreen-ui/modal';
 export { uiColors } from '@leafygreen-ui/palette';
-export * as compassUIColors from './components/compass-ui-colors';
+export * as compassUIColors from './compass-ui-colors';
 export { default as Portal } from '@leafygreen-ui/portal';
 export { Select, Option, Size as SelectSize } from '@leafygreen-ui/select';
 export { Tabs, Tab } from '@leafygreen-ui/tabs';
